exclude duplicate tests
