+++ acceptance-tests/test/end-to-end/tests/BARMultisite_test.js
@@ -6,9 +6,6 @@ const testConfig = require('../config/BARConfig');
 let paymentReferenceSite1 = '';
 let paymentReferenceSite2 = '';
 
-const toCheckXPath = '//div/app-card[2]/div/div[1]/h1';
-const submittedXPath = '//div/app-card[1]/div/div[1]/h1';
-
 function createPayment(name, amount, ref, that) {
   that.waitForElement('#payment_type_POSTAL_ORDER', BARATConstants.fiveSecondWaitTime);
   that.click('#payment_type_POSTAL_ORDER');
@@ -86,125 +83,6 @@ Before(({ I }) => {
   };
 });
 
-Scenario('@functional Fee-clerk switches sites and check payment visibility', async({ I }) => {
-  await I.login(testConfig.TestBarSwitchSiteUserName, testConfig.TestBarSwitchSitePassword);
-  I.wait(BARATConstants.tenSecondWaitTime);
-  await I.switchSite('MILTON KEYNES COUNTY COURT');
-  const payerNameSite1 = await generatePayerName(I);
-  I.clickAddPayment();
-  createPayment(payerNameSite1, '273', paymentReferenceSite1, I);
-  await I.switchSite('MEDWAY COUNTY COURT');
-  I.wait(BARATConstants.tenSecondWaitTime);
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  const payerNameSite2 = await generatePayerName(I);
-  I.clickAddPayment();
-  createPayment(payerNameSite2, '273', paymentReferenceSite2, I);
-  I.clickPaymentList();
-  I.dontSee(payerNameSite1);
-  await I.switchSite('MILTON KEYNES COUNTY COURT');
-  I.wait(BARATConstants.tenSecondWaitTime);
-  I.dontSee(payerNameSite2);
-  await I.Logout();
-}).retry(testConfig.ScenarioRetryLimit);
-
-// Not needed to run in the pipeline as it's covered in the diff scenarios
-Scenario('Fee-clerk "check and submit" page validate with multisite', async({ I }) => {
-  await I.login(testConfig.TestBarSwitchSiteUserName, testConfig.TestBarSwitchSitePassword);
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  await I.switchSite('MILTON KEYNES COUNTY COURT');
-  I.clickCheckAndSubmit();
-  I.wait(BARATConstants.twoSecondWaitTime);
-  const payerNameSite1 = await generatePayerName(I);
-  const toCheckSite1 = await I.grabTextFrom(toCheckXPath);
-  const submittedSite1 = await I.grabTextFrom(submittedXPath);
-
-  // Check on the other site
-  I.switchSite('MEDWAY COUNTY COURT');
-  I.wait(BARATConstants.twoSecondWaitTime);
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  I.clickCheckAndSubmit();
-  I.wait(BARATConstants.twoSecondWaitTime);
-  const payerNameSite2 = await generatePayerName(I);
-  const toCheckSite2 = await I.grabTextFrom(toCheckXPath);
-  I.say(`Found ${toCheckSite2} payments in XPath ${toCheckXPath}`);
-  const submittedSite2 = await I.grabTextFrom(submittedXPath);
-
-  // Process a payment instruction and send to "check and submit"
-  I.processPayment();
-  I.clickCheckAndSubmit();
-  I.retry(BARATConstants.retryCountForStep).see(parseInt(toCheckSite2) + 1, toCheckXPath);
-  I.see(payerNameSite2);
-
-  // We check that on the other site nothing has changed
-  I.switchSite('MILTON KEYNES COUNTY COURT');
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  I.clickCheckAndSubmit();
-  I.say(`Expecting to see: ${toCheckSite1} in: ${toCheckXPath}`);
-  I.see(toCheckSite1, toCheckXPath);
-  I.wait(BARATConstants.twoSecondWaitTime);
-  I.dontSee(payerNameSite2);
-  I.wait(BARATConstants.twoSecondWaitTime);
-  // Switch back site to submit payment
-  I.switchSite('MEDWAY COUNTY COURT');
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  I.checkAndSubmit(payerNameSite2, 'Submit');
-  const numberToCheck = parseInt(submittedSite2) + 1;
-
-  I.waitForText(numberToCheck, BARATConstants.tenSecondWaitTime);
-  // even though it's found the text sometimes it hasn't fully loaded :(
-  I.wait(BARATConstants.twoSecondWaitTime);
-  I.see(numberToCheck, submittedXPath);
-  I.say(`Expecting to see: ${toCheckSite2} in: ${toCheckXPath}`);
-  I.waitForText(toCheckSite2, BARATConstants.tenSecondWaitTime);
-  I.see(toCheckSite2, toCheckXPath);
-  pause();
-  // Switch back to validate that the numbers are correct on the other site
-  I.switchSite('MILTON KEYNES COUNTY COURT');
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  I.clickCheckAndSubmit();
-  I.waitForText(toCheckSite1, BARATConstants.tenSecondWaitTime);
-  I.say(toCheckSite1);
-  I.say(toCheckXPath);
-
-  I.retry(BARATConstants.retryCountForStep).see(toCheckSite1, toCheckXPath);
-  I.see(submittedSite1, submittedXPath);
-
-  // process other sites payments
-  I.processPayment();
-  I.clickCheckAndSubmit();
-  I.see(parseInt(toCheckSite1) + 1, toCheckXPath);
-  I.see(payerNameSite1);
-
-  // Validate we don't see the payment on the other site (site 2)
-  I.switchSite('MEDWAY COUNTY COURT');
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  I.clickCheckAndSubmit();
-  I.dontSee(payerNameSite1);
-  // submit the newly processed payment on site1
-  I.switchSite('MILTON KEYNES COUNTY COURT');
-  I.clickCheckAndSubmit();
-  I.checkAndSubmit(payerNameSite1, 'Submit');
-  const expectedNumber = parseInt(submittedSite1) + 1;
-  I.waitForText(expectedNumber, BARATConstants.tenSecondWaitTime);
-  // even though it's found the text sometimes it hasn't fully loaded :(
-  I.wait(BARATConstants.twoSecondWaitTime);
-  I.see(expectedNumber, submittedXPath);
-  I.see(toCheckSite1, toCheckXPath);
-
-  // and finally check again on the other site
-  I.switchSite('MEDWAY COUNTY COURT');
-  I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
-  I.clickCheckAndSubmit();
-  I.waitForText(toCheckSite2, BARATConstants.tenSecondWaitTime);
-  // even though it's found the text sometimes it hasn't fully loaded :(
-  I.wait(BARATConstants.twoSecondWaitTime);
-  I.see(numberToCheck, submittedXPath);
-  I.see(toCheckSite2, toCheckXPath);
-
-  // Logout
-  await I.Logout();
-}).retry(testConfig.ScenarioRetryLimit);
-
 Scenario('@functional Fee-clerk "check and submit" validate action counter', async({ I }) => {
   await I.login(testConfig.TestBarSwitchSiteUserName, testConfig.TestBarSwitchSitePassword);
   I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
@@ -227,9 +105,9 @@ Scenario('@functional Fee-clerk "check and submit" validate action counter', asy
   I.waitForText('Add payment', BARATConstants.tenSecondWaitTime);
   const payerNameSite2 = await generatePayerName(I);
   I.clickAddPayment();
-  createPayment(payerNameSite2, '273', paymentReferenceSite1, I);
+  createPayment(payerNameSite2, '273', paymentReferenceSite2, I);
   I.click('Add another payment');
-  createPayment(payerNameSite2, '273', paymentReferenceSite1, I);
+  createPayment(payerNameSite2, '273', paymentReferenceSite2, I);
 
   // Collect action info
   const site2Before = await collectActionNumbers(I);
@@ -306,7 +184,7 @@ Scenario('@functional Fee-clerk Advance search for multi site -  All statuses', 
   await I.switchSite('MEDWAY COUNTY COURT');
   const payerNameSite2 = await generatePayerName(I);
   I.clickAddPayment();
-  createPayment(payerNameSite2, '273', paymentReferenceSite1, I);
+  createPayment(payerNameSite2, '273', paymentReferenceSite2, I);
   I.click('Advanced search');
   I.wait(BARATConstants.tenSecondWaitTime);
   I.waitForText('Action', BARATConstants.twoSecondWaitTime);
